add multithread for cosmos auto update

+++ src/wgpu/wcommand_buffer.cpp
@@ -102,7 +102,7 @@ void RenderPassCommand::set_pipe_line(PipeLinePtr pipe_line)
 void RenderPassCommand::set_vertex_buffer(size_t slot, BufferPtr buffer)
 {
 	size_t size = wgpuBufferGetSize(buffer->buffer_);
-	wgpuRenderPassEncoderSetVertexBuffer(render_pass_encoder_, slot, buffer->buffer_, 0, size);
+	wgpuRenderPassEncoderSetVertexBuffer(render_pass_encoder_, uint32_t(slot), buffer->buffer_, 0, size);
 }
 
 void RenderPassCommand::set_bind_group(size_t slot, BindGroupPtr bind_group)
@@ -112,7 +112,8 @@ void RenderPassCommand::set_bind_group(size_t slot, BindGroupPtr bind_group)
 
 void RenderPassCommand::draw(size_t vertex, size_t instance, size_t first_vertex, size_t first_instance)
 {
-	wgpuRenderPassEncoderDraw(render_pass_encoder_, vertex, instance, first_vertex, first_instance);
+	wgpuRenderPassEncoderDraw(render_pass_encoder_, 
+		(uint32_t)vertex, (uint32_t)instance, (uint32_t)first_vertex, (uint32_t)first_instance);
 }
 
 void RenderPassCommand::submit()
